requirements.txt
boto3>=1.26
python-dotenv
certifi
tqdm